% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftt_mvdf.R
\name{as.data.frame,ftt_mvdf-method}
\alias{as.data.frame,ftt_mvdf-method}
\title{Coerce to a data frame}
\usage{
\S4method{as.data.frame}{ftt_mvdf}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object inheriting from \code{ftt_mvdf}.}

\item{row.names, optional, ...}{Arguments passed to \link[base:as.data.frame]{base::as.data.frame}.}
}
\description{
Coerce any object inheriting from \code{ftt_mvdf} to a data frame.
}
