% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forthetrees-package.R
\docType{package}
\name{forthetrees-package}
\alias{forthetrees}
\alias{forthetrees-package}
\title{forthetrees: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikemahoney218/forthetrees}
  \item \url{https://mikemahoney218.github.io/forthetrees/}
  \item Report bugs at \url{https://github.com/mikemahoney218/forthetrees/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
