% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftt_add_cakepop.R
\name{ftt_add_cakepop}
\alias{ftt_add_cakepop}
\alias{ftt_add_cakepop_method,mvdf_obj-method}
\alias{ftt_add_cakepop_method,ftt_treesize-method}
\title{Add code to add primitives to a Blender script}
\usage{
ftt_add_cakepop(
  script,
  object,
  height = NULL,
  dbh = NULL,
  crown = TRUE,
  crown_radius = NULL,
  ...
)

\S4method{ftt_add_cakepop_method}{mvdf_obj}(
  object,
  script,
  height = NULL,
  dbh = NULL,
  crown = TRUE,
  crown_radius = NULL,
  ...
)

\S4method{ftt_add_cakepop_method}{ftt_treesize}(
  object,
  script,
  height = NULL,
  dbh = NULL,
  crown = TRUE,
  crown_radius = NULL,
  ...
)
}
\arguments{
\item{script}{The Python script to append the generated code onto.}

\item{object}{An object inheriting from \link{mvdf_obj} which will be used to
calculate the x, y, and z positions for each primitive. One primitive will be
created for each row of \code{mvdf(object)}.}

\item{height}{The height of the central bole, in meters.}

\item{dbh}{The diameter of the central bole, in meters. Cake pops do not
taper.}

\item{crown}{Boolean: draw a "crown" sphere?}

\item{crown_radius}{The radius of the crown, in meters. The crown is centered
on the top of the bole.}

\item{...}{Additional arguments to pass to the primitive creation call. The
available arguments are different for each primitive, and are documented in
the official Blender documentation at
\url{https://docs.blender.org/api/blender_python_api_current/bpy.ops.mesh.html}.}
}
\value{
A length 1 character vector containing the Blender Python script with
code for creating primitives added.
}
\description{
This function generates code that will, when run inside Blender, create mesh
primitives within a scene.
}
