% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftt_add_sapling.R
\name{ftt_add_sapling}
\alias{ftt_add_sapling}
\title{Add code to add primitives to a Blender script}
\usage{
ftt_add_sapling(
  script,
  object,
  id_code = NULL,
  id_key = NULL,
  location = NULL,
  ...
)
}
\arguments{
\item{script}{The Python script to append the generated code onto.}

\item{object}{An object inheriting from \link{mvdf_obj} which will be used to
calculate the x, y, and z positions for each primitive. One primitive will be
created for each row of \code{mvdf(object)}.}

\item{id_code}{x}

\item{id_key}{x}

\item{location}{x}

\item{...}{Additional arguments to pass to the sapling creation call.
Currently ignored.}
}
\value{
A length 1 character vector containing the Blender Python script with
code for creating saplings added.
}
\description{
This function generates code that will, when run inside Blender, create
saplings (using the "add curve sapling" addon) to a scene.
}
