% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftt_add_simple_trees.R
\name{ftt_add_simple_trees}
\alias{ftt_add_simple_trees}
\alias{ftt_add_simple_trees,mvdf_obj-method}
\title{Add code to add simple trees to a Blender script}
\usage{
ftt_add_simple_trees(object, script, ...)

\S4method{ftt_add_simple_trees}{mvdf_obj}(object, script, ...)
}
\arguments{
\item{object}{An object inheriting from \link{mvdf_obj} which will be used to
calculate the x, y, and z positions for each primitive. One primitive will be
created for each row of \code{mvdf(object)}.}

\item{script}{The Python script to append the generated code onto.}

\item{...}{Ignored}
}
\value{
A length 1 character vector containing the Blender Python script with
code for creating trees added.
}
\description{
This function generates code that will, when run inside Blender, create
simple trees using the sapling tree gen add-on.
}
