% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{ftt_contains}
\alias{ftt_contains}
\title{Test to ensure an object meets minimal content requirements}
\usage{
ftt_contains(object, class, error = FALSE)
}
\arguments{
\item{object}{The object to test.}

\item{class}{The S4 class (as a character string) with the slots that
\code{object} needs to have.}

\item{error}{Logical: Should the function error if the object doesn't contain
the other class, or return \code{FALSE}? Default to \code{FALSE} (do not error).}
}
\description{
Test to ensure an object meets minimal content requirements
}
