% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftt_mvdf.R
\name{appendix}
\alias{appendix}
\alias{appendix,ftt_mvdf-method}
\title{Retrieve the appendix from a \code{forthetrees} object}
\usage{
appendix(object)

\S4method{appendix}{ftt_mvdf}(object)
}
\arguments{
\item{object}{The object to retrieve the appendix from.}
}
\description{
Retrieve the appendix from a \code{forthetrees} object
}
