% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftt_mvdf.R
\name{metadata}
\alias{metadata}
\alias{metadata,ftt_mvdf-method}
\title{Retrieve the metadata data frame from a \code{forthetrees} object}
\usage{
metadata(object)

\S4method{metadata}{ftt_mvdf}(object)
}
\arguments{
\item{object}{The object to retrieve the metadata data frame from.}
}
\description{
Retrieve the metadata data frame from a \code{forthetrees} object
}
