% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftt_mvdf.R
\docType{class}
\name{ftt_mvdf-class}
\alias{ftt_mvdf-class}
\title{The Minimum Viable Data Frame S4 class underpinning \code{forthetrees}}
\description{
The Minimum Viable Data Frame S4 class underpinning \code{forthetrees}
}
\section{Slots}{

\describe{
\item{\code{idx}}{Character: a unique identifier (or "index") for each object to be
modeled. Must be unique with no \code{NA} or \code{NULL} values, but otherwise is
not validated.}

\item{\code{x,y,z}}{Numeric: distance of the origin of the object from the origin of
the grid system (the central point at 0, 0, 0) in meters in the given
direction. Must have no \code{NA}, \code{NULL}, \code{NaN}, \code{Inf}, or \code{-Inf} values.}

\item{\code{metadata}}{Data frame: a table containing additional information on the
objects to be modeled. Optional, but if this slot is used then the data frame
must contain a column named \code{idx} which should correspond to the \code{idx} slot.
Only the existence of this column is validated.}

\item{\code{appendix}}{List: additional data produced in the generation of the
object. Not validated; any additional outputs that don't map to modeled
objects may be inserted here.}
}}

\section{Units}{

For cohesiveness, \code{forthetrees} assumes all values are in SI units. In
practice, this means:
\itemize{
\item All linear units are in meters.
\item All measurements of area are in m^2
\item All measurements of volume are in m^3
}

It is the responsibility of importer functions to correctly convert from
non-standard units into these formats.
}

\seealso{
Other classes and related functions: 
\code{\link{ftt_tree-class}},
\code{\link{ftt_treesize-class}}
}
\concept{classes and related functions}
