% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftt_mvdf.R
\name{ftt_mvdf}
\alias{ftt_mvdf}
\title{Construct a \code{forthetrees} Minimum Viable Data Frame object}
\usage{
ftt_mvdf(data, metadata = data.frame(), appendix = list())
}
\arguments{
\item{data}{Data frame containing columns with names corresponding to the
slots listed in TK.}

\item{metadata}{Data frame: a table containing additional information on the
objects to be modeled. Optional, but if this slot is used then the data frame
must contain a column named \code{idx} which should correspond to the \code{idx} slot.
Only the existence of this column is validated.}

\item{appendix}{List: additional data produced in the generation of the
object. Not validated; any additional outputs that don't map to modeled
objects may be inserted here.}
}
\description{
Construct a \code{forthetrees} Minimum Viable Data Frame object
}
