% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftt_mvdf.R
\name{ftt_set_mvdf}
\alias{ftt_set_mvdf}
\title{Set slot values for \code{forthetrees} objects.}
\usage{
ftt_set_mvdf(mvdf, object, metadata = NULL, appendix = NULL)
}
\arguments{
\item{mvdf}{The minimum viable data frame required by the S4 class}

\item{object}{The}

\item{metadata}{The metadata to include in the new object. If \code{NULL}
(the default), uses the metadata from \code{object}.}

\item{appendix}{The appendix to include in the new object. If \code{NULL}
(the default), uses the appendix from \code{object}.}
}
\description{
This function returns a new object of the same class as \code{object} with the
same metadata and appendix slots as \code{object}, but with other slots taking
values from \code{mvdf}. Use it as a convenient, pipe-able way to set slot values
for \code{forthetrees} objects.
}
