% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{ftt_treesize-class}
\alias{ftt_treesize-class}
\title{S4 class for tree sizes: diameter, height, and crown specifics.}
\description{
S4 class for tree sizes: diameter, height, and crown specifics.
}
\section{Slots}{

\describe{
\item{\code{dbh}}{Numeric: diameter at breast height, in meters. May include missing
values, which will be handled differently by different rendering functions.}

\item{\code{height}}{Numeric: bole height, in meters. May include missing values,
which will be handled differently by different rendering functions. Height is
understood as the total height of the object to be rendered (top of the stem
and crown, if applicable); different importer functions may determine this
differently.}

\item{\code{crown}}{Logical: does the tree have leaves? May include missing values,
which will be handled differently by different rendering functions.}

\item{\code{crown_radius}}{Numeric: radius of the crown, in meters, at its widest
point. May include missing values, which will be handled differently by
different rendering functions.}

\item{\code{idx}}{Character: a unique identifier (or "index") for each object to be
modeled. Must be unique with no \code{NA} or \code{NULL} values, but otherwise is
not validated.}

\item{\code{x,y,z}}{Numeric: distance of the origin of the object from the origin of
the grid system (the central point at 0, 0, 0) in meters in the given
direction. Must have no \code{NA}, \code{NULL}, \code{NaN}, \code{Inf}, or \code{-Inf} values.}

\item{\code{metadata}}{Data frame: a table containing additional information on the
objects to be modeled. Optional, but if this slot is used then the data frame
must contain a column named \code{idx} which should correspond to the \code{idx} slot.
Only the existence of this column is validated.}

\item{\code{appendix}}{List: additional data produced in the generation of the
object. Not validated; any additional outputs that don't map to modeled
objects may be inserted here.}
}}

\seealso{
Other classes and related functions: 
\code{\link{ftt_mvdf-class}},
\code{\link{ftt_tree-class}}
}
\concept{classes and related functions}
