% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{ftt_treesize}
\alias{ftt_treesize}
\title{Create an \code{ftt_treesize} object}
\usage{
ftt_treesize(
  data = NULL,
  dbh = "dbh",
  height = "height",
  crown = "crown",
  crown_radius = "crown_radius",
  ...
)
}
\arguments{
\item{data}{Optionally, a data frame containing all the data necessary to
create an \code{ftt_treesize} object. If \code{NULL}, all other arguments are
interpreted as data to use in constructing the object; if not \code{NULL},
arguments are interpreted as the names of columns in \code{data} containing the
values for each slot.}

\item{dbh}{Numeric: diameter at breast height, in meters. May include missing
values, which will be handled differently by different rendering functions.}

\item{height}{Numeric: bole height, in meters. May include missing values,
which will be handled differently by different rendering functions. Height is
understood as the total height of the object to be rendered (top of the stem
and crown, if applicable); different importer functions may determine this
differently.}

\item{crown}{Logical: does the tree have leaves? May include missing values,
which will be handled differently by different rendering functions.}

\item{crown_radius}{Numeric: radius of the crown, in meters, at its widest
point. May include missing values, which will be handled differently by
different rendering functions.}

\item{...}{Additional arguments passed to \link[mvdf:mvdf_simple_material]{mvdf::mvdf_simple_material}.}

\item{translate_colors}{Boolean passed to \link[mvdf:mvdf_simple_material]{mvdf::mvdf_simple_material}.}
}
\value{
An object of class \code{ftt_treesize}.
}
\description{
Create an \code{ftt_treesize} object
}
