% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{ftt_treespp}
\alias{ftt_treespp}
\title{Create an \code{ftt_treespp} object}
\usage{
ftt_treespp(data = NULL, id_code = "id_code", id_key = "id_key", ...)
}
\arguments{
\item{data}{Optionally, a data frame containing all the data necessary to
create an \code{ftt_treespp} object. If \code{NULL}, all other arguments are
interpreted as data to use in constructing the object; if not \code{NULL},
arguments are interpreted as the names of columns in \code{data} containing the
values for each slot.}

\item{id_code}{The code indicating what sort of identifier is stored in
id_key -- one of fia_code, plants_code, common_name, genus, or species.}

\item{id_key}{The key identifying the tree species in the schema referenced
by id_code.}

\item{...}{Additional arguments passed to \link[mvdf:mvdf_obj]{mvdf::mvdf_obj}.}
}
\value{
An object of class \code{ftt_treespp}.
}
\description{
Create an \code{ftt_treespp} object
}
