% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{move_location}
\alias{move_location}
\title{Build the location commands for classes inheriting from \code{mvdf_obj}}
\usage{
move_location(mvdf_df)
}
\arguments{
\item{object}{An object inheriting from \code{mvdf_obj}}
}
\description{
Build the location commands for classes inheriting from \code{mvdf_obj}
}
\seealso{
Other utilities: 
\code{\link{pythonize_booleans}()}
}
\concept{utilities}
\keyword{internal}
