% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftt_mvdf.R
\name{mvdf}
\alias{mvdf}
\alias{mvdf,ftt_mvdf-method}
\title{Retrieve mvdf values from a \code{forthetrees} object.}
\usage{
mvdf(object)

\S4method{mvdf}{ftt_mvdf}(object)
}
\arguments{
\item{object}{The \code{forthetrees} object to retrieve the mvdf for.}
}
\description{
Retrieve mvdf values from a \code{forthetrees} object.
}
