% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftt_mvdf.R
\name{mvdf<-}
\alias{mvdf<-}
\alias{mvdf<-,ftt_mvdf-method}
\title{Set mvdf values for a \code{forthetrees} object.}
\usage{
mvdf(x) <- value

\S4method{mvdf}{ftt_mvdf}(x) <- value
}
\arguments{
\item{x}{The \code{forthetrees} object to set the mvdf within.}

\item{value}{The data to replace the mvdf with.}
}
\description{
Set mvdf values for a \code{forthetrees} object.
}
