% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pythonize_booleans}
\alias{pythonize_booleans}
\title{Convert R logical values to Python equivalents}
\usage{
pythonize_booleans(dots)
}
\arguments{
\item{dots}{The results of running \code{list(...)} in an exporter function.}
}
\description{
Convert R logical values to Python equivalents
}
\seealso{
Other utilities: 
\code{\link{move_location}()}
}
\concept{utilities}
\keyword{internal}
