% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_lookups.R
\name{ftt_fia_sapling_lookup}
\alias{ftt_fia_sapling_lookup}
\alias{ftt_plants_sapling_lookup}
\alias{ftt_common_name_sapling_lookup}
\alias{ftt_species_sapling_lookup}
\alias{ftt_genus_sapling_lookup}
\title{Retrieve sapling rendering methods for a given lookup value.}
\usage{
ftt_fia_sapling_lookup(fia_code)

ftt_plants_sapling_lookup(plants_code)

ftt_common_name_sapling_lookup(common_name)

ftt_species_sapling_lookup(species)

ftt_genus_sapling_lookup(genus)
}
\arguments{
\item{fia_code}{The FIA species code (See appendix F of Burrill et al. 2018)
to search for.}

\item{plants_code}{The USDA PLANTS species code (USDA 2021)
to search for.}

\item{common_name}{The common name used to identify a particular species
(See appendix F of Burrill et al. 2018) to search for.}

\item{species}{The species to lookup, as a single string in the format
\verb{genus species} (See appendix F of Burrill et al. 2018) to search for.}

\item{genus}{The genus to search for.}
}
\value{
The Python command used to generate renders for the given lookup.
}
\description{
This function retrieves the Python command used to render given species of
trees, as identified by their FIA identifier, USDA PLANTS code identifier,
common name (as specified by the FIA program), species, or genus. Note that
capitalization is ignored when matching all character-based lookups -- all
lookup values are stored as lower case and all queries are coerced to lower
case. If no render method has been defined for the lookup value, a default
method is returned.
}
\references{
Burrill, Elizabeth A.; Wilson, Andrea M.; Turner, Jeffery A.; Pugh, Scott A.;
Menlove, James; Christiansen, Glenn; Conkling, Barbara L.; David, Winnie.
2018. The Forest Inventory and Analysis Database: database description and
user guide version 8.0 for Phase 2. U.S. Department of Agriculture, Forest
Service. 946 p. (Online). Available at web address:
\url{http://www.fia.fs.fed.us/library/database-documentation}.

USDA, NRCS. 2021. The PLANTS Database (\url{http://plants.usda.gov},
14 January 2021). National Plant Data Team, Greensboro, NC 27401-4901 USA.
}
