% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftt_mvdf.R
\name{show,ftt_mvdf-method}
\alias{show,ftt_mvdf-method}
\title{Show an object inheriting from \code{ftt_mvdf}}
\usage{
\S4method{show}{ftt_mvdf}(object)
}
\arguments{
\item{object}{Any object inheriting from \code{ftt_mvdf}}
}
\value{
\code{show} returns an invisible \code{NULL}.
}
\description{
Display the object, by printing, plotting or whatever suits its class.
}
