% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{solve_tree}
\alias{solve_tree}
\title{Find approximately reasonable default values for tree size parameters}
\usage{
solve_tree(dbh = NA, height = NA, crown_radius = NA, region = "nw")
}
\arguments{
\item{dbh}{Diameter at breast height, in meters.}

\item{height}{Bole height, in meters.}

\item{crown_radius}{Crown radius, in meters.}

\item{region}{Currently ignored. Exists to allow future expansion of
parameters, to let functions more realistically scale trees for a given
region.}
}
\value{
A named list with default values for dbh, height, and crown_radius.
}
\description{
Find approximately reasonable default values for tree size parameters
}
\references{
B. R. Lockhart, R. C. Weih, and K. M. Smith. 2005 Crown Radius and Diameter
at Breast Height Relationships for Six Bottomland Hardwood Species. Journal
of the Arkansas Academy of Science. 59.
Available at:http://scholarworks.uark.edu/jaas/vol59/iss1/16

L. Zhang. 1997. Cross-validation of Non-linear Growth Functions for
Modelling Tree Height–Diameter Relationships. Annals of Botany. 79(3):
251-257. https://doi.org/10.1006/anbo.1996.0334.

R. C. Yang, A. Kozak, and J. H. G. Smith. 1978. The potential of
Weibull-type functions as flexible growth curves. Canadian Journal of Forest
Research. 8(4): 424-431. https://doi.org/10.1139/x78-062
}
