% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_positions}
\alias{tree_positions}
\title{Tree positions}
\format{
A data frame with 1279 observations of 4 values:
\describe{
\item{EAST}{Eastings, in meters from the lower-left corner}
\item{NORTH}{Northings, in meters from the lower-left corner}
\item{SPECIES}{A numeric code indicating what species a tree is}
\item{DBH85}{Diameter at breast height, in inches}
}
}
\usage{
tree_positions
}
\description{
Tree positions
}
\keyword{datasets}
